requirements.txt
python-dotenv==1.0.1
langchain==0.2.2
langchain-community==0.2.3
langchain-huggingface==0.0.3
unstructured==0.14.4
chromadb==0.5.0
tiktoken==0.7.0
streamlit==1.35.0
huggingface-hub==0.23.2

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.8

# Sentence transformers for embeddings
sentence-transformers==2.7.0

# For Hugging Face Transformers (fallback)
transformers==4.41.2
torch==2.3.1
accelerate==0.30.1
datasets==2.19.1

# For Ollama integration
langchain-ollama==0.1.0
requests==2.31.0